requirements.txt
# Required packages for Botany-VQA dataset generation

# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
pillow>=9.0.0
pandas>=1.5.0
numpy>=1.23.0

# Computer vision
opencv-python>=4.7.0
scikit-learn>=1.2.0
scikit-image>=0.20.0

# Dataset utilities
datasets>=2.12.0
tqdm>=4.65.0

# Jupyter notebook
jupyter>=1.0.0
ipywidgets>=8.0.0
matplotlib>=3.7.0

# Optional: For faster processing
accelerate>=0.20.0
bitsandbytes>=0.39.0  # For 8-bit quantization
